baisc char showing the amounts billed is in place
